requirements.txt
asgiref==3.5.2
coverage==6.4.4
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
Pillow==9.2.0
python-dotenv==0.21.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
